// src/enum.ts
export enum Mode {
  View = 'View',
  Edit = 'Edit'
}

export enum Mouse {
  Mousedown = 'mousedown',
  Mousemove = 'mousemove',
  Mouseup = 'mouseup'
}

export enum Key {
  Keydown = 'keydown',
  Keyup = 'keyup'
}

export enum ShapeType {
  Rect = 'Rect',
  Ellipse = 'Ellipse',
  Polygon = 'Polygon'
}

export enum FreeDrawEvents {
  Transform = 'Transform',
  Drag = 'Drag'
}

export enum KeyCodes {
  Backspace = 'Backspace',
  Control = 'Control',
  Enter = 'Enter'
}